dd tests for track call
