+++ packages/browser-destinations/src/destinations/logrocket/__tests__/index.test.ts
@@ -1,46 +1,15 @@
 import { Analytics, Context } from '@segment/analytics-next'
 import plugins, { destination } from '../index'
-import { Subscription } from '../../../lib/browser-destinations'
-// import WorkerStub from 'test/worker-stub'
-
-const subscriptions: Subscription[] = [
-  {
-    partnerAction: 'track',
-    name: 'Track',
-    enabled: true,
-    subscribe: 'type = "track"',
-    mapping: {
-      name: {
-        '@path': '$.name'
-      },
-      properties: {
-        '@path': '$.properties'
-      }
-    }
-  }
-]
-
-class WorkerStub {
-  url: string
-  onmessage: (_arg: string) => void
-  constructor(stringUrl: string) {
-    this.url = stringUrl
-    this.onmessage = (_arg: string) => {}
-  }
-
-  postMessage(msg: string) {
-    this.onmessage(msg)
-  }
-
-  addEventListener() {}
-}
+import { subscriptions } from './subscriptions'
+import { mockWorkerAndXMLHttpRequest } from './utilities'
 
 describe('Logrocket', () => {
+  beforeAll(mockWorkerAndXMLHttpRequest)
+  afterAll(jest.restoreAllMocks)
+
   test('can load', async () => {
     const [event] = await plugins({ appID: 'log/rocket', subscriptions })
 
-    window.XMLHttpRequest = jest.fn()
-    window.Worker = WorkerStub
     jest.spyOn(destination, 'initialize')
 
     await event.load(Context.system(), {} as Analytics)
