+++ packages/browser-destinations/src/destinations/logrocket/__tests__/subscriptions.ts
@@ -1,0 +1,18 @@
+import { Subscription } from '../../../lib/browser-destinations'
+
+export const subscriptions: Subscription[] = [
+  {
+    partnerAction: 'track',
+    name: 'Track',
+    enabled: true,
+    subscribe: 'type = "track"',
+    mapping: {
+      name: {
+        '@path': '$.name'
+      },
+      properties: {
+        '@path': '$.properties'
+      }
+    }
+  }
+]
