+++ packages/browser-destinations/src/destinations/logrocket/index.ts
@@ -38,6 +38,7 @@ export const destination: BrowserDestinationDefinition<Settings, LR> = {
   initialize: async ({ settings: { appID } }, deps) => {
     LogRocket.init(appID)
     await deps.resolveWhen(() => Object.prototype.hasOwnProperty.call(window, '_LRLogger'), 100)
+    window.LogRocket = LogRocket // not functionally required, but helpful for tests/debug
     return LogRocket
   },
