+++ packages/browser-destinations/src/destinations/logrocket/track/__tests__/index.test.ts
@@ -1,6 +1,30 @@
-// import { Analytics, Context } from '@segment/analytics-next'
-// import { destination } from '../../index'
+import { Analytics, Context } from '@segment/analytics-next'
+import plugins from '../../index'
+import { subscriptions } from '../../__tests__/subscriptions'
+import { mockWorkerAndXMLHttpRequest } from '../../__tests__/utilities'
 
-// describe('Logrocket.track', () => {
-//   // TODO: Test your action
-// })
+describe('Logrocket.track', () => {
+  beforeAll(mockWorkerAndXMLHttpRequest)
+  afterAll(jest.restoreAllMocks)
+  it('sends track events to logrocket', async () => {
+    const [event] = await plugins({ appID: 'log/rocket', subscriptions })
+
+    await event.load(Context.system(), {} as Analytics)
+    const logRocket = jest.spyOn(window.LogRocket, 'track')
+
+    const name = 'testName'
+    const properties = {
+      goodbye: 'moon'
+    }
+
+    await event.track?.(
+      new Context({
+        type: 'track',
+        name,
+        properties
+      })
+    )
+
+    expect(logRocket).toHaveBeenCalledWith(name, properties)
+  })
+})
